requirements.txt
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.3

# Async web scraping
aiohttp==3.9.1
asyncio==3.4.3

# Environment management
python-dotenv==1.0.0

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
requests==2.31.0